backend/requirements.txt
Flask == 2.0.1
Flask-Cors == 3.0.10
Pillow == 9.0.1
matplotlib == 3.6.2
numpy==1.23.5
opencv-python==4.6.0.66
scikit-learn==1.1.3
scipy==1.9.3
tensorflow==2.11.0
wrapt==1.14.1
zipp==3.10.0
imutils==0.5.4